+++ crates/persist-pty/src/process.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+pub struct ProcessId(pub u32);
+
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+pub struct ProcessGroupId(pub u32);